requirements.txt
Django==3.2.11
psycopg2-binary==2.8.6
gunicorn
django-filter==21.1
djangorestframework==3.13.1
drf-yasg==1.20.0
channels==3.0.4
django-cors-headers==3.11.0
djangorestframework-jwt==1.11.0
